Change `TimeEntry.Time` to a `float64`
I had forgotten that time spent can be a decimal. Make this field a
float and change related code accordingly:

* --time flag
* `NewTimeEntry()` `time` argument
* `buildSubmissionParams()` can't use `strconv.Itoa` as that's an
  integer function. Instead we use `FormatFloat`. Truncate time parsing
  to two decimal places because to me that seems like enough. Who's
  going to say they spent `0.324` hours on something? Also, in order to
  be able to properly submit values to TimeTask (at least on the edition
  I use), the times must be written in European/French format, with
  commas (`,`) as decimal separators. Do a string replace to get this,
  as the float formatter will give us a period (`.`) separator.

+++ main.go
@@ -40,7 +40,7 @@ func main() {
 	time_spent := kingpin.Flag("time", "Time spent working on project.").
 		Short('t').
 		Default("7").
-		Int()
+		Float()
 	date_str := kingpin.Flag("date", "Date when work was done (e.g. 2017-01-31)").
 		Short('d').
 		String()
